Tests/LibWeb/Text/expected/css/animation-in-shadow-dom.txt

animations: 1
at 0.25: none
at 0.5: none
